LPD-52196 Move tests
