+++ modules/test/playwright/tests/fragment-web/fragmentAdminMarketplace.spec.ts
@@ -12,6 +12,7 @@ import {isolatedSiteTest} from '../../fixtures/isolatedSiteTest';
 import {loginTest} from '../../fixtures/loginTest';
 import {pageEditorPagesTest} from '../../fixtures/pageEditorPagesTest';
 import {pageManagementSiteTest} from '../../fixtures/pageManagementSiteTest';
+import getRandomString from '../../utils/getRandomString';
 import performLogin, {performLogout, userData} from '../../utils/performLogin';
 
 const test = mergeTests(
@@ -126,5 +127,72 @@ test(
 				.getByRole('dialog')
 				.getByRole('heading', {name: 'Marketplace is now in'})
 		).toBeVisible();
+	}
+);
+
+test(
+	'Check available actions of marketplace fragment',
+	{
+		tag: '@LPD-34938',
+	},
+	async ({apiHelpers, fragmentsPage, page, site}) => {
+
+		// Create new fragment collection
+
+		const fragmentCollectionName = getRandomString();
+
+		const {fragmentCollectionId} =
+			await apiHelpers.jsonWebServicesFragmentCollection.addFragmentCollection(
+				{
+					groupId: site.id,
+					name: fragmentCollectionName,
+				}
+			);
+
+		const fragmentName = getRandomString();
+
+		await apiHelpers.jsonWebServicesFragmentEntry.addFragmentEntry({
+			fragmentCollectionId,
+			groupId: site.id,
+			html: `<div class="fragment-example">
+				  Example marketplace fragment
+				</div>`,
+			marketplace: true,
+			name: fragmentName,
+			type: 'component',
+		});
+
+		// Go to fragment administration
+
+		await fragmentsPage.goto(site.friendlyUrlPath);
+
+		// Click the More Actions button to open the actions
+
+		await page
+			.locator('.card-row')
+			.filter({hasText: fragmentName})
+			.getByLabel('More actions')
+			.click();
+
+		// Check available actions
+
+		['View Usages', 'Move', 'Delete'].forEach(async (action) => {
+			await expect(
+				page.getByRole('menuitem', {name: action})
+			).toBeVisible();
+		});
+
+		[
+			'Edit',
+			'Change Thumbnail',
+			'Mark as Cacheable',
+			'Export',
+			'Make a Copy',
+			'Rename',
+		].forEach(async (action) => {
+			await expect(
+				page.getByRole('menuitem', {name: action})
+			).not.toBeVisible();
+		});
 	}
 );
