+++ modules/test/playwright/tests/fragment-web/fragmentsAdmin.spec.ts
@@ -1863,72 +1863,3 @@ testEmbeddingWidgets(
 		).toBeVisible();
 	}
 );
-
-test.describe('Marketplace Fragments', () => {
-	test(
-		'Check available actions of marketplace fragment',
-		{
-			tag: '@LPD-34938',
-		},
-		async ({apiHelpers, fragmentsPage, page, site}) => {
-
-			// Create new fragment collection
-
-			const fragmentCollectionName = getRandomString();
-
-			const {fragmentCollectionId} =
-				await apiHelpers.jsonWebServicesFragmentCollection.addFragmentCollection(
-					{
-						groupId: site.id,
-						name: fragmentCollectionName,
-					}
-				);
-
-			const fragmentName = getRandomString();
-
-			await apiHelpers.jsonWebServicesFragmentEntry.addFragmentEntry({
-				fragmentCollectionId,
-				groupId: site.id,
-				html: `<div class="fragment-example">
-				  Example marketplace fragment
-				</div>`,
-				marketplace: true,
-				name: fragmentName,
-				type: 'component',
-			});
-
-			// Go to fragment administration
-
-			await fragmentsPage.goto(site.friendlyUrlPath);
-
-			// Click the More Actions button to open the actions
-
-			await page
-				.locator('.card-row')
-				.filter({hasText: fragmentName})
-				.getByLabel('More actions')
-				.click();
-
-			// Check available actions
-
-			['View Usages', 'Move', 'Delete'].forEach(async (action) => {
-				await expect(
-					page.getByRole('menuitem', {name: action})
-				).toBeVisible();
-			});
-
-			[
-				'Edit',
-				'Change Thumbnail',
-				'Mark as Cacheable',
-				'Export',
-				'Make a Copy',
-				'Rename',
-			].forEach(async (action) => {
-				await expect(
-					page.getByRole('menuitem', {name: action})
-				).not.toBeVisible();
-			});
-		}
-	);
-});
